converting tests to tap; removing KeyValue
