--- src/KeyValue.ts
@@ -1,8 +1,0 @@
-export class KeyValue {
-  readonly key: any;
-  readonly item: any;
-  constructor(key, item) {
-    this.key = key;
-    this.item = item;
-  }
-}
