+++ src/utils/typeTests.ts
@@ -12,7 +12,6 @@ import {
   isUndefined,
 } from './tests';
 
-// @ts-ignore
 export const TESTS = [
   { name: TypeEnum.undefined, test: isUndefined, isForm: false },
   { name: FormEnum.map, test: isMap, isForm: true },
